Convert all tabs to PDF containers
- Replace HTML content components with PDF viewers for all tabs
- Create Tournament Rules PDF with comprehensive tournament policies
- Create Code of Conduct PDF with community guidelines and enforcement
- Update PDFViewer component to dynamically load different PDFs based on active tab
- Update MobilePDFViewer component with same PDF functionality
- Maintain all PDF viewer controls (search, zoom, navigation, download) for all tabs
- Remove HTML Tournament Rules and Code of Conduct components
- Ensure consistent user experience across all rule sections

Features:
- Dynamic PDF loading based on tab selection
- Appropriate filenames for downloads per tab
- Reset page navigation when switching tabs
- Unified PDF viewer interface for all content
- Mobile-responsive design maintained
- Professional PDF documents with proper formatting
