+++ public/assets/konivrer-code-of-conduct.pdf
@@ -1,0 +1,137 @@
+%PDF-1.4
+%“Œ‹ž ReportLab Generated PDF document http://www.reportlab.com
+1 0 obj
+<<
+/F1 2 0 R /F2 3 0 R /F3 4 0 R
+>>
+endobj
+2 0 obj
+<<
+/BaseFont /Helvetica /Encoding /WinAnsiEncoding /Name /F1 /Subtype /Type1 /Type /Font
+>>
+endobj
+3 0 obj
+<<
+/BaseFont /Helvetica-Bold /Encoding /WinAnsiEncoding /Name /F2 /Subtype /Type1 /Type /Font
+>>
+endobj
+4 0 obj
+<<
+/BaseFont /Helvetica-BoldOblique /Encoding /WinAnsiEncoding /Name /F3 /Subtype /Type1 /Type /Font
+>>
+endobj
+5 0 obj
+<<
+/Contents 12 0 R /MediaBox [ 0 0 612 792 ] /Parent 11 0 R /Resources <<
+/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
+>> /Rotate 0 /Trans <<
+
+>> 
+  /Type /Page
+>>
+endobj
+6 0 obj
+<<
+/Contents 13 0 R /MediaBox [ 0 0 612 792 ] /Parent 11 0 R /Resources <<
+/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
+>> /Rotate 0 /Trans <<
+
+>> 
+  /Type /Page
+>>
+endobj
+7 0 obj
+<<
+/Contents 14 0 R /MediaBox [ 0 0 612 792 ] /Parent 11 0 R /Resources <<
+/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
+>> /Rotate 0 /Trans <<
+
+>> 
+  /Type /Page
+>>
+endobj
+8 0 obj
+<<
+/Contents 15 0 R /MediaBox [ 0 0 612 792 ] /Parent 11 0 R /Resources <<
+/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
+>> /Rotate 0 /Trans <<
+
+>> 
+  /Type /Page
+>>
+endobj
+9 0 obj
+<<
+/PageMode /UseNone /Pages 11 0 R /Type /Catalog
+>>
+endobj
+10 0 obj
+<<
+/Author (\(anonymous\)) /CreationDate (D:20250704013137+00'00') /Creator (\(unspecified\)) /Keywords () /ModDate (D:20250704013137+00'00') /Producer (ReportLab PDF Library - www.reportlab.com) 
+  /Subject (\(unspecified\)) /Title (\(anonymous\)) /Trapped /False
+>>
+endobj
+11 0 obj
+<<
+/Count 4 /Kids [ 5 0 R 6 0 R 7 0 R 8 0 R ] /Type /Pages
+>>
+endobj
+12 0 obj
+<<
+/Filter [ /ASCII85Decode /FlateDecode ] /Length 1486
+>>
+stream
+Gatm:_/A9k&A@6Wk*QpEC.bc_+.M'h%9g,QlKtPe>,!<\,uQ%gAGu9"86L,)G_O^f$q,gqF5FQTo4AianR:C+d1Y9<l%aUT/>i*+=;n$Cre9dk=8uO1rL;!#Z;ljMEA(dB[`TFU0k8CC5uEfAbBZRS]#$c"E:sa$CN4;eEU"'p@t/<\7Tl&;n[bO^T7?b2iQnF6-i0K/6$eHi.f7)28*eL.rgn"6^C<HCj7@1cB`Ri"0$?b`=e<T*MUFIe`%-_NQ,ik6.Q]oOEtFuk8X/2g)lV1_.8!`^-VAEJ$7!'EWZkLPg'I##X--YG02`NZDrn<uY2QuI.5jsHT\5qf^3%4_TaG4BWt"TrToL'>$e]O]CinTe--KteV;EV0OjD(TQRGG(ciO@.RNqjn?/!?#Z]K=tZbC#`b$Y?8)W6rf,O>]pAC*cV%iPu67S3@WkOW8V^(37HJgM&_+Lcbq7SgmEU)<ZP(oYnu1,dl74:h2CV7.:.6.tU,ea"=70MH>E=ufqZg5;cT0;2XffGT,*b=^]-*T(aI!Em4ZSMld?UoVTW;3+h_$f)aLln5#bC^T",]ol-bc9W6XHn3n[_Kb%+:l2H/A1<`NDp;t5Mn^d;HWAeQn`[],^%%.H)"".%1a'q1j>.b$jJ\a'<jA!j<)+8?S0#'**Hago4'A6F$uI7S.d[r#Bmua2a)oY(TuU4DEAHApSgl<b:6tTnW7t-G0]!*`W7D;]g$AcK>9+(;]<("o+O`^8iJBqH'Q,?qN@mW`=[$eFrLk:dI"l^36(0UU**5,P3RN#qdp-mlW)m2q,;5[_04"tR):f[.GSE-MQ5b[d9G8I[QQ\Jb5<<YOD7Qs[m&"5=X7ZFUIP[$mrp,mqq\K$C9"8C><B:5WUe5F[3NmDKAd>g;^0*af;YRmWBVWsPcQGiLTjFSJK(joqk93YRRa1q2M&0U?Au.t9UP?3*Zi]'QIdgoCS2tN7<H-r!A`aITh:9Aua\!h:lk6`F7rZt_oQ2Tur[bIF%%PNp+b:K]q@bgCg*(8$o245YaB#>!*0V5:/XDrD_nXiEM3+YtXAKZhr3PY$SYBK6@q]60L9%=EMbIPEp1IkpUqP_&Zti*,S![&!<%6$"mbPr6oWf3=HJZ;Xe:8?eKZ8n<Hh51O[[8"t"#4,RMtXStRU*-7Qo-9AO!n9#a-0>+c^XqWSTo-,Lem;)1pBaA/G7:tc"TL'q8iXW;E])>I=>^8Dm<L)\a;ZH,b"KP_^!1jpnWM"GTP`.MC/MPV[`%J1u*Mq65ET\DQ;QK$c@GG#D+S>%W_g>_c2#LT,[eTFGW#<RU]m\E@3b?R)<U*N1QRZI9dfjifc").IFALd(CDhfN?M1@T%h04ZE:kiZ6$LB;G9`d3:[]O<:p'aggloUe>M"/rYXlaj>=g%Em39K4;Ue%;`E,(%jDL<IPhE-T(65_8#FJn&q]?bPZ2ASR4X1?UK&i#SC/QXhX$h[9o?+$bNmZUT:SkJC?O~>endstream
+endobj
+13 0 obj
+<<
+/Filter [ /ASCII85Decode /FlateDecode ] /Length 900
+>>
+stream
+Gatn%9lldX&;KZL'mo8e_*rkT+.G*$B:@&?XaBA&L,E!_,U/)Tm]Z'd*61W_FO7r72J01bGB_qMR6a.E)rKrki-Yj'Xr+AWQ>9n(/;N=E>V=@jce.3uZ@Xi'l!_/iN;$"m7R(D'Po)'aBb`P/P57&ms']IF^TC<aFPL];!H!?[LP@EsDArBQQ?TS2k]&iNG_RcL=QHm@>7]g*2POn^JP^-!3pqZ%YS1oL#7!D\:0:/uJguX['c1&r_*/JtpUsV+SFoIX4k>WTD8I13A(g$7XO-;sC<'O84rYc#1/O0t_:?-70KR<'oSf&(?8@nECgL'EI9(&?cX)H"-S['S@MDb#@^6m,9"mdt5LEC,<igrg:5%SI3[]0`8D&Neg%^/%Ss37S)P0XlN9iWop]+rG1IccV`;P53@?<OR,.%6u+?J0@_&TNB^#q9Zeo=&a<!,[h?3>5r1Y*_J-&^;o+d!_S-`#1CWc=3@CN"V_$Ls)\\6b@$_c:*L)SRHi'=Xn!"#n<CM(Mfg+#Udai'am^Z1CC8q\q+?nQ:QjFrE29&(2c@DF4L,s7/o5h\K1f/h!QTQ1EUG:PK:ZOCf&UHl>a570[>Ie$nRuPQk(<%8,.13".RKf_P->8%s]Rkf(PsbPU/[h?`N\*jcou?p!L%(@LchW?s_d@XX&R._?QD>+f#+&,fmh%B@JbdGpil#orAfVEK0!+ZAD2_oZqRI)culgO[Q[<)CRaQptrq)`!h(<!T<s/U*A3e[Q\2.k0c4XYM:T0Aa^0);Lsk>&d-R<6\7O+>teqZZ3m:EcaLO@K2/?ls##jCkX,MrPbG;Vd-6U2lPUV*`*;*D%Le,N'4:P"cNQZ)!_O*<\<"M5#(,][ts-Pd9e$042a,OSrZk9RUet'HTCWSlp'o?[#G~>endstream
+endobj
+14 0 obj
+<<
+/Filter [ /ASCII85Decode /FlateDecode ] /Length 1407
+>>
+stream
+Gatm;bAu>q']&X:mP=1>/75.4V44:JFIHq0C9m+/-:%qe#gdh;<PH0PJ4(oe9?h/2[j2mTnt*n8.O`>M&V%4kd1:$hrl9VJP1Dc2F]Quf=*hWqH+GF7="[S(gblSYcXk)$WL!d"A[,oL%=YM'C/L,Hgu+i73ZFt+34G]XM<D%YWGh'\rZ$4LR+$e=ntP/B-uQenod<pm)c6RMkEj951*PiN\Z^"1KS0,ZLf6d5^P_-9IXAHWlo5Hu[9IOj$H\QsH[6HFX1]aO*&g#>6>^<b=Bo%G#6kH]',sH:7jFD]e-YH.C,trV%l?ne#<jl[fmFmA9<3<88SL\n@#'IMZh@K%87:7Ki%M!<-/?_r:g@2A"!=pjAA%!1'3[.=['B`+POCCPa"Y!aL>8)o`641A1V>:,0!E8!'cq61SIms7mYRp'fM7?b`iXE1A(\:#g`5uBE\duC:TN0S(A0PdZH27ClEt@%!7f?rQE(*qH']=$C0&:ab3FAbWrEHo-bp(,:S'cN('g;F4b&tmRY.bEoc+.si`-KVeXf57S5tSB3I9QIcsm30Kb>M#rXW0a3$`6q^$(/l%tP!ZD`)Wq#B.$84F"6+7s;p>hj>ZIJZiM??#Ufh&$<+4[K,9n2i<=?ih.?X0U^>%Pa)1#U*)RCKJV9<>S3,eBQ%GraTNm=Na!kh!9f$LL(Ws=br0BX-*GC0o9MW:fVM`lZC/*lodWHbr$+H#k_f$g,UPZ1'9$XDBnQ,Ae:G!]'XH14ADEA];P::O\X7MY3-"`Ih+esn'@t'[b"c%2hKa1P0HA(%C)8.P$c32cY8H5]?$hc=YYWt4$(#oUo)(O*<-1?bSB4n<Wb-AS"2C`rNp\U6jT;1cX^KQ`Sk*9%3X@h&q9PRRoLmPT&or:5=*j1aO9aA,'8sd3m(H)U%a6\K@L,^HQr;_`2R'4PCP5rVp&UO6ar06]S.D/?YHL5uK=q((Pb&tBjdGf%Fme*VnOA[pC8FF*SHPN1YAgCQ_LX0!+E>[SY-naDTBltA*"tl7p_uLAO!D<8+l:.S)I7?Lk^ll?NA2@2c/>3i(Q`mUd.7C,F`X*#QZV,nH5LaJT%PlkD>+!gP6Yo!>-UT!(@2F-.Dpf1$+'96j2?cq6Ado5.>"108M[^D%)_%d)9e]!:i0NZ.W-X-TN/sVB/;S01l\-[E/&(K3a0a^.IfE''F`R[in=RR!k\-VT(qit$PQ%DS`L#u]+-J/IH8QBn"TiDe+A1HCmLWdSfRN;@Ot-`f2CmH@\@9KdHqN/CkahpOl77/(d2$%Og)"^4[iG*UgkJC=<%?:T6,&!]=;,c5IqDl!b2<OY7Dd9bo[c-3ugK6jSULrG*X#Z'-#n.\qNji%Wlj`q]$@-7%MI113\iWJ@.H&.LjZWY?um2b(6aRU76W5LFh@p~>endstream
+endobj
+15 0 obj
+<<
+/Filter [ /ASCII85Decode /FlateDecode ] /Length 730
+>>
+stream
+Gat=h9okN(&;KZF/*c83'PRai`oELF1$Yb`d>O>t.%o1h)nVT-nF*iTq:IQoN!\H!1fQWNoi<f'Bk-bFl2e%\Gl*$m9;$*OV,7T]B=KBlIol"3AmR_D-hK/WMs"`]?.!<O*/Qu7+/5u"8ePT<rFU<b[g"m%Kl]Cl&&s4N@Ml0aX:u_L1>7KeWM/U1LM#&$QU\TR:u4S8.Zf%Um2m6Q/>_%Z@&sTo<MsCS'M.CVFPmVqUX`hNFC$d5&7e"M2-hKd91ci'WBD7FfptGB^n]HDK-D4H%A4lTN)ih$i/YT*Misk@P_eE`bo&nUbBUkf=sWEMr<t;=^a6ek$s1Yl70N'RhiCt;$*HdtC8osGH/j.C"M[c5VFT"CLI$b'XTP84kYM`H47#8,Ykh8gXB:qh#ItK+O#D?=Fc1<Mk:LtbQ"1.MbYsuloU^07=Fc>`Abd[SQ^&OWVU,T]r=;o#fkTcOdCIUB0Gl1Ngu2g_B'E27l4T"-FMarZ[i^*;DmiH/G!_0YJWr4J<rB^De)K^hnmmnNP`dJZ%7GRWZOMEPcLbr!jH5Mc;jMB+/o.uN@ZUI.g3;VR(KAVodimOD2l")Gdj)Rc/WulF"P4<pPu;3ijU<JWGf"on7m&DaZ.i.,(cp!VL-:MFMkN+^mXo/i\#=N!)EeUirLB1RALBji=p[njDhhbiSY@pcU\3a`5#dd+$qNZ'+g%3-1Bu%Sd@))*nu*k9*WTBnLq*~>endstream
+endobj
+xref
+0 16
+0000000000 65535 f 
+0000000073 00000 n 
+0000000124 00000 n 
+0000000231 00000 n 
+0000000343 00000 n 
+0000000462 00000 n 
+0000000657 00000 n 
+0000000852 00000 n 
+0000001047 00000 n 
+0000001242 00000 n 
+0000001311 00000 n 
+0000001595 00000 n 
+0000001673 00000 n 
+0000003251 00000 n 
+0000004242 00000 n 
+0000005741 00000 n 
+trailer
+<<
+/ID 
+[<61be9c58905fc60dbc14dbccadaaba5a><61be9c58905fc60dbc14dbccadaaba5a>]
+% ReportLab generated PDF document -- digest (http://www.reportlab.com)
+
+/Info 10 0 R
+/Root 9 0 R
+/Size 16
+>>
+startxref
+6562
+%%EOF
