+++ public/assets/konivrer-tournament-rules.pdf
@@ -1,0 +1,137 @@
+%PDF-1.4
+%“Œ‹ž ReportLab Generated PDF document http://www.reportlab.com
+1 0 obj
+<<
+/F1 2 0 R /F2 3 0 R /F3 4 0 R
+>>
+endobj
+2 0 obj
+<<
+/BaseFont /Helvetica /Encoding /WinAnsiEncoding /Name /F1 /Subtype /Type1 /Type /Font
+>>
+endobj
+3 0 obj
+<<
+/BaseFont /Helvetica-Bold /Encoding /WinAnsiEncoding /Name /F2 /Subtype /Type1 /Type /Font
+>>
+endobj
+4 0 obj
+<<
+/BaseFont /Helvetica-BoldOblique /Encoding /WinAnsiEncoding /Name /F3 /Subtype /Type1 /Type /Font
+>>
+endobj
+5 0 obj
+<<
+/Contents 12 0 R /MediaBox [ 0 0 612 792 ] /Parent 11 0 R /Resources <<
+/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
+>> /Rotate 0 /Trans <<
+
+>> 
+  /Type /Page
+>>
+endobj
+6 0 obj
+<<
+/Contents 13 0 R /MediaBox [ 0 0 612 792 ] /Parent 11 0 R /Resources <<
+/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
+>> /Rotate 0 /Trans <<
+
+>> 
+  /Type /Page
+>>
+endobj
+7 0 obj
+<<
+/Contents 14 0 R /MediaBox [ 0 0 612 792 ] /Parent 11 0 R /Resources <<
+/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
+>> /Rotate 0 /Trans <<
+
+>> 
+  /Type /Page
+>>
+endobj
+8 0 obj
+<<
+/Contents 15 0 R /MediaBox [ 0 0 612 792 ] /Parent 11 0 R /Resources <<
+/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
+>> /Rotate 0 /Trans <<
+
+>> 
+  /Type /Page
+>>
+endobj
+9 0 obj
+<<
+/PageMode /UseNone /Pages 11 0 R /Type /Catalog
+>>
+endobj
+10 0 obj
+<<
+/Author (\(anonymous\)) /CreationDate (D:20250704013055+00'00') /Creator (\(unspecified\)) /Keywords () /ModDate (D:20250704013055+00'00') /Producer (ReportLab PDF Library - www.reportlab.com) 
+  /Subject (\(unspecified\)) /Title (\(anonymous\)) /Trapped /False
+>>
+endobj
+11 0 obj
+<<
+/Count 4 /Kids [ 5 0 R 6 0 R 7 0 R 8 0 R ] /Type /Pages
+>>
+endobj
+12 0 obj
+<<
+/Filter [ /ASCII85Decode /FlateDecode ] /Length 1399
+>>
+stream
+Gatn&9lo&K%#46H'f`cd[4COcp)_s_B/?$cfu14p>\FGLfh/&,;GOm>IR-U*WZ@qrFZaXVP6)f<!!.g?5VM9WZO5XN$pK&N^d=D'^uA3"ZWjr2>s2JuDWc6Gkqo3\o\#$(Cmm*rKoD#?"@c6,O+S/=iee/LXmI:"X2o5c@I+DXTei/`UZmG4B/pD"qY'UETDRr$R&NPJ?D4`Fhiqg2_<a8-IY6V9DW:J]?AD4ZBDVGQ+Pr3?$-eY\kp3*tP;#V',LS.l9>(B?#<H7fIBEgUS'3c/2f,C:')HK/LJoQ5C#cs9'F3YiI-`k,4\);oX*s0!mCSsll^]:Kh*t\cLJ]AC3:j0Ii0\Qtpm5=&i"J$Y0B%4G50WKP^97UQD#h$pUK(fh7^-)6"4+h:6"A#D0F*hrdM[@)/orG'CFS<G$C*J\["09'=7q^q@6dC4)YkF.D_c3aU0l_tRB!8V$G2%<,KcN-N8Q$Q![.@_nJUVTY>LdkLTTe])q8&&A6@m;o1?KEn:Y$tH:';beS'XO.st@'HX'B<7/]NrMF':1+5W/^26E7BW55,Bf]d(N(7>>#7DB:F:X]pR.2GS+[#b6[B<hsKak/d1(6\T+gda$s`RZ+;;4oY8AQ[q!9DIB)$R]_^0QAj.UQ4d9S@trHS]6QJ33K'R$d`!h.91@!@iL=6CKF+eE[@e?SY;5.)Q5h\[G-9OC4'cLG/_#G\Z,@^W1'`o#,\A8V<q,t[SAYkMNi8WqA"q*b8g9g^2TPnJ$$u6XRj"p_G?HPqUgV]Ao@+Z`aef-UaD3K=ATVkD,bB,j;F*?>]K/V+3`+r,Kb1>+&+RTQfbX$K[ZN/aO-`bKVZP#b)<8J\*i`Is1ip]`odEY.QWM`O':R9W5c4$ZDF9!;J2skSUnYA$;$L=2cO.egUo+3E\:Et$dV/^6Xf-CNhk.J.^t8sf@qibHC7*Zi8IfU"*\!`=!eGnA$P<A*Md>j5Y\!f'KOc\d&N5=J];(&:j2oZ"4/;KNW4@N9.joelo1(k]PqmL?8n#e$2NTh/u1R8e6D":o,*@Xj;M8rd1cV%9!+/OQRY.blDrW<7jjnsUj4Ic.p](7cuhmm-J5KDH!,<+]kHrHB]#c9<7GPk;pF'<f9%&=34BZ7mZQ4\&POfsb7!R#-g8e3P#f"MDDjLtb@:@dOW=m$+*TXQ2ktGuT7u<&R05cQ1mT%V9e'/W(Xs`gVZ_79Mlp43OH>^4_R_qh,/BP6lEEL+oMQjBg/g`5C(SuM/r,ZeO&gRHiGgO]L=:!aO7.Btj%HY+g:r%NF<\j@^gB(W4V8UsJE\alIq(gdKul0ceO3Ma.G-j=qa%?B)Mu'3OP=9pBNH<o5T73\'6-bX^3>pOrO#m/Jb4M3Rc"-,7+&ic9)#>J.RU`B%r)R6K`~>endstream
+endobj
+13 0 obj
+<<
+/Filter [ /ASCII85Decode /FlateDecode ] /Length 1172
+>>
+stream
+Gatn%gN)%,&;KZP'RQ6k<*%pCUbC2omU;8*D)@HGgd'q`7SEk?%SE:icN9mfgX/8"<RQgq"&\qK^0UXE#"s9Kn6$8%^T?DW)[7jIFUK`rE`g)2hptm'=)<>kXIrhpLM8>3YF"U%"\BJ9OQ=TjatqE#'a8'-@]p#oRo.*QY)iE'9,Oaqduk4i/\KUSQf8D<+(5$Rej"K6J*?i!HF9b"l"*tD0#Y%eM0?[d;AteVDGg#>#MaX!JP4jnf3CSE3ao+gblYCkTG<Xa^n&/=>1a'V')_dr+j5F@MqTGg,[^J-AlR379omXN%4'nO!?-Aoc)jf_!McNOJT9_K[%4!6`AI6>2Qdk%9;6dZe&u,pi=FuYr8j@WUOm<-#4.i>%(5EKJ<30;e7\q$'1e'QJ"R#cXG$t]NRJ$j&Y]BqpMNC4'o$2*]cu=*r%=j3.#l^J^%V+Qet_o<Zqc=SnO]B`q^niG14V0j$+U5V-7W/oB-;YuUp(Q*RZUcdVPr$G1cTR=`Y7:97\U/ucQB*"G?eo@N+dpK5QWU8q2BWKA<6QCTp[F<;9hMQFYabnf8%Lk<#dPr;u74&3*`nEbK<gk1.Y;DX=Agl<bYT^$TFWbScW,MX]JrsGlk1R^4TcA#HM-$-753ha*[S0hiFii%gptaf"<dm3+IAX=Y:uXImg#\J:&6m7&uiEk9f_j9"Ik@^1tKD^6l!h"T>;tT.)eh3r>(XEH8U75/k)."SP!dekBT6?#;A'lR?\4;SgE![NQHOnPlf8;gCs314e(hdUPemj)N`+-Y^p?*L1:V-`erK-GDI:8"tIR`(AoF_]a#?i;*qQR!_r3QT,sR!q1`/MR_t_q<AS^=3psI^pPF`M5or<rc`YCT>nCO\Ha;BF6>>Y@anSDS^Cms[';hu5s?Q)]/->fr`54CM.*2;fT:e)RN''\pElfl<n)`\8OZ\\N7q?:<L6";_XB[nf?-KSPb,DKS76(u7m/88^)MZ.U&'&uda&h%jSm@eU%U)I!P0!@n`nFsmBlQlaKpic\]im*T[EST&%0l6deM`_aTo4Mc90!"3'',Gd0e,c*!jIp/;-EQd=q*0`SsHo"A96.KXl<>%lHL-o[rGYSiF;]OGN6=jBfj`=f6%fChOe@+Q0"F2*UK14Xsm=_ZmOAiNsLc[/HscEiY](!*<M.[gcS/~>endstream
+endobj
+14 0 obj
+<<
+/Filter [ /ASCII85Decode /FlateDecode ] /Length 1192
+>>
+stream
+Gatm;99\?n&AHJokZ^Rf%^M'K\P;d[7jU$V%(3Y47G)IPM(NuuR47j]V^5j*g"Fme1Umc?>P.[_f2!7-l!jY*BE<Qg_iO7(YSbq+YV\s\r<Ku>?ThZ!_oZI"fd`c7F>%+H/6t#e@^aAKN'?hm=X9[+.l,&'[`M]GG%4pJ\!oI[ZRg^oY!_%_\CITEA%m`k4fj-PJ:Gu>'Qb3'TQ>VO9*dF@&9[+Y@Lr!k/XLG2n3(Lg/`-[)J8jD42+!.R0LA>5<e9'Nr//.?hqbU]PJpKJUUtA]/8UCbPmF+$YnYu&6hDjdg/HCI,@hgjO\DTkEq;A6'ENl<8k@eA1"iBh0O*'d-"q,*ORusB/8Wiq;"g=t4p8qlpMr_SQK`1\iA-"m^l:h_0+eTo0'38oF/K7E:TfcT^l1_?D0fIeD"]6c`>8Frs)1aN-nVXmVo,4U^l3"pZs<(9*m$^E1u(5$6S_gt9@>?72jqH?Ia7\bl]`Z`E$"O2/<h]X;@h2d'5/`=C&YtRI$*2&[2%kG932e:V#^nK,5L*GTo6/j`eoA^T?29C9UAaI*l6`9'=9(fe/9[bi*@RBYn;u"q]LMN^U-YGhP`S"Ac)`]kL"KAX*Q?h4m?-_/S$R.e+lra+*]#&&1-:kMR<*b\C!V@>2qt'1mYie$e9E6KbZZPgsPX7Td67&kdkrQ'F[i@5HG?9N5gRL5mcp^kt.I#O6fjJ6t-^>X]0R^7IV_lp@B[]?#qGY^!K`BYt85eZ`D/)bYJL_?/CH5n$EG]cEjK:3#i%uOmA0N@D`Pf;j-q?q7-K<2<d*8L^/Lp^9Ba8$("Q_Z>c$<<V(\i-I/LZ+N5FKWgZNkFEbe=,#Fu;^mg'j\J$M"SmEa1.88>t$47N*C3E@(4fhHIhi+NO,Vf"Ob;,4),IjMN7SkkQ&\@11QNor]!1&\:4'p/LmFP4p?P'(*AY)*+Se'H[/a:0Q5%6=XCXf,_ef`n\bBu\IjK"G.klG[SZbmR!NKS\<+EY7]:mFh/#_%ug[8H%`R:G>r]4JTt5NCjcqZ06gk=B_bHf(.ON27)sZ"*XT4Y;KX2]\&!RY)poqA#p3,\.<B]eqVPi7]#&`%ZkW$EqH&#YM[!jT](CDP6JD,ZIsTi7DLW'cj-BY@o!BA#UJ,kLM4tUs?3tT+lFF8_'C'K`4tJOVA8Va4WodA[DD7;n2Z7D?-_2~>endstream
+endobj
+15 0 obj
+<<
+/Filter [ /ASCII85Decode /FlateDecode ] /Length 294
+>>
+stream
+Gar'#_/=lZ%#445`II5taC7D"YLD4SG*nLW!"E8V"JI-8;Vup-3c/X`j^A&d5+A4AB_cKU4<$EM$'`;SJo[-o]OGN:F$d?H=eX>d=F)IQC>'G"g%(eV>ul6*nX'T<rif*Mge<MSGG]7KWslV#D&`fL/7/,2/;FIt0:9f5qC_He>SP_'ek>Jg6eNrP")&:CMs,";\;kQD#L<FYD_T+q8+<^o=l&;>&92(\)=Wq21t1[d&Ub[:b<b)3n?)GkkGG]8>m0'8=34`&EgXEDkAas<GL=Ofh,-_W!TaPrnG~>endstream
+endobj
+xref
+0 16
+0000000000 65535 f 
+0000000073 00000 n 
+0000000124 00000 n 
+0000000231 00000 n 
+0000000343 00000 n 
+0000000462 00000 n 
+0000000657 00000 n 
+0000000852 00000 n 
+0000001047 00000 n 
+0000001242 00000 n 
+0000001311 00000 n 
+0000001595 00000 n 
+0000001673 00000 n 
+0000003164 00000 n 
+0000004428 00000 n 
+0000005712 00000 n 
+trailer
+<<
+/ID 
+[<dfbc75ec3e91acc33ab17cd6a39cd8d6><dfbc75ec3e91acc33ab17cd6a39cd8d6>]
+% ReportLab generated PDF document -- digest (http://www.reportlab.com)
+
+/Info 10 0 R
+/Root 9 0 R
+/Size 16
+>>
+startxref
+6097
+%%EOF
